requirements.txt
fastapi==0.95.2
uvicorn==0.24.0
pydantic==1.10.12
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
PyJWT==2.8.0

